Moved extension methods to a single location.
