+++ AO3-Statistics.sln
@@ -14,8 +14,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AO3Statistics.ConsoleApp", "AO3Statistics.Console\AO3Statistics.ConsoleApp.csproj", "{34FF5E47-9F5E-4D21-92E5-FDBF3CD08D0D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StringExtensions", "StringExtensions\StringExtensions.csproj", "{37234791-149C-48D5-BC32-93EA5E2BF9B6}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -48,14 +46,6 @@ Global
 		{34FF5E47-9F5E-4D21-92E5-FDBF3CD08D0D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{34FF5E47-9F5E-4D21-92E5-FDBF3CD08D0D}.Release|x64.ActiveCfg = Release|Any CPU
 		{34FF5E47-9F5E-4D21-92E5-FDBF3CD08D0D}.Release|x64.Build.0 = Release|Any CPU
-		{37234791-149C-48D5-BC32-93EA5E2BF9B6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{37234791-149C-48D5-BC32-93EA5E2BF9B6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{37234791-149C-48D5-BC32-93EA5E2BF9B6}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{37234791-149C-48D5-BC32-93EA5E2BF9B6}.Debug|x64.Build.0 = Debug|Any CPU
-		{37234791-149C-48D5-BC32-93EA5E2BF9B6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{37234791-149C-48D5-BC32-93EA5E2BF9B6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{37234791-149C-48D5-BC32-93EA5E2BF9B6}.Release|x64.ActiveCfg = Release|Any CPU
-		{37234791-149C-48D5-BC32-93EA5E2BF9B6}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
